refactor: fix mistake in folder name
