+++ src/components/Chip/Chip.stories.tsx
@@ -2,7 +2,6 @@ import type { Meta, StoryObj } from '@storybook/react';
 
 import { Chip } from './Chip';
 
-// More on how to set up stories at: https://storybook.js.org/docs/writing-stories#default-export
 const meta = {
   component: Chip,
   tags: ['autodocs'],
